requirements.txt
# Example packages for Python 3.12 compatibility
numpy>=1.25.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.30.0
flask>=2.3.0
contextualized-topic-models>=2.6.0

# Video generation dependencies (security-patched versions)
moviepy>=1.0.3
Pillow>=10.2.0  # CVE fixes for libwebp and arbitrary code execution
aiohttp>=3.13.3  # CVE fixes for zip bomb, DOS, and directory traversal
pydantic>=2.5.0